Moved files into a neat structure
