+++ Program/header_files/CoverImage.h
