+++ Program/header_files/CoverPixel.h
