+++ Program/object_files/CoverPixel.cpp
@@ -1,12 +1,11 @@
-#include "CoverPixel.h"
-#include "Converter.h"
-
+#include "../header_files/CoverPixel.h"
+#include "../header_files/Converter.h"
+#include "../header_files/sha512.h"
 #include <iostream>
 #include <string.h>
 #include <stdlib.h>
-#include "sha512.h"
 #include <sodium.h>
-#include "vector"
+#include <vector>
 
 using namespace std;
