+++ Program/object_files/FrankDecode.cpp
@@ -1,11 +1,11 @@
-#include "FrankDecode.h"
-#include "CoverImage.h"
-#include "Converter.h"
-#include "CoverPixel.h"
-#include "ImageToOutput.h"
-#include "FileToOutput.h"
-#include "ImageToRead.h"
-#include "FileToDecode.h"
+#include "../header_files/FrankDecode.h"
+#include "../header_files/CoverImage.h"
+#include "../header_files/Converter.h"
+#include "../header_files/CoverPixel.h"
+#include "../header_files/ImageToOutput.h"
+#include "../header_files/FileToOutput.h"
+#include "../header_files/ImageToRead.h"
+#include "../header_files/FileToDecode.h"
 #include <string>
 #include <iostream>
 #include <vector>
